requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
requests
sqlmodel
httpx
openai
pgvector
sqlalchemy
numpy
ddgs
streamlit
urllib3
pandas
tabulate
langsmith
kiwipiepy
redis
elasticsearch==8.11.0
arize-phoenix
opentelemetry-exporter-otlp
openinference-instrumentation-openai
redis-om